// src/infra/definitions/crops.type.ts
export enum CROPY_TYPES {
  SOYBEAN = 'SOYBEAN',
  CORN = 'CORN',
  WHEAT = 'WHEAT',
  RICE = 'RICE',
  SORGHUM = 'SORGHUM',
  BARLEY = 'BARLEY',
  OATS = 'OATS',
  MILLET = 'MILLET',
  BEAN = 'BEAN',
  PEANUT = 'PEANUT',
  CHICKPEA = 'CHICKPEA',
  LENTIL = 'LENTIL',
  SESAME = 'SESAME',
  MANIOC = 'CASSAVA',
  SWEET_POTATO = 'SWEET POTATO',
  POTATO = 'POTATO',
  YAM = 'YAM',
  TARO = 'TARO',
  ORANGE = 'ORANGE',
  LEMON = 'LEMON',
  PINEAPPLE = 'PINEAPPLE',
  MANGO = 'MANGO',
  PASSION_FRUIT = 'PASSION FRUIT',
  BANANA = 'BANANA',
  PAPAYA = 'PAPAYA',
  GRAPE = 'GRAPE',
  WATERMELON = 'WATERMELON',
  MELON = 'MELON',
  COFFEE = 'COFFEE',
  SUGAR_CANE = 'SUGARCANE',
  EUCALYPTUS = 'EUCALYPTUS',
  PINE = 'PINE',
  MATE = 'MATE',
  COCOA = 'COCOA',
  LETTUCE = 'LETTUCE',
  TOMATO = 'TOMATO',
  ONION = 'ONION',
  CARROT = 'CARROT',
  CABBAGE = 'CABBAGE',
  PEPPER = 'BELL PEPPER',
  KALE = 'KALE',
  PUMPKIN = 'PUMPKIN',
  CUCUMBER = 'CUCUMBER',
  COTTON = 'COTTON',
  TOBACCO = 'TOBACCO',
  SUNFLOWER = 'SUNFLOWER',
  CASTOR_BEAN = 'CASTOR BEAN',
  PASTURE = 'PASTURE',
}